backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
neo4j==5.17.0
networkx==3.2.1
pandas==2.2.0
faker==23.2.1
groq
pydantic-settings==2.2.1
python-dotenv==1.0.1
scikit-learn==1.4.0
numpy==1.26.4
python-multipart==0.0.9
shap==0.44.1
anthropic==0.21.0
websockets==12.0
